# swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Student-Supervisor Relationships",
        "title": "SSR Application",
        "contact": {
            "name": "Ivakhnenko Dmitry, Moscow",
            "url": "github.com/ivaahn/ssr",
            "email": "ivahnencko01@gmail.com"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login into account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User email",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/student/bid": {
            "get": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "GetUserInfo student's bids",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "student_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StBids"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Apply bid",
                "parameters": [
                    {
                        "description": "bid info",
                        "name": "ApplyBid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ApplyBid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ApplyBidResp"
                        }
                    }
                }
            }
        },
        "/api/student/feedback": {
            "get": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Get feedbacks on the supervisor.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Supervisor ID",
                        "name": "supervisor_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FeedbackPlenty"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Provide a feedback",
                "parameters": [
                    {
                        "description": "feedback info",
                        "name": "Feedback",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FeedbackReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.FeedbackAddResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/student/profile": {
            "get": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "GetUserInfo student's profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StProfile"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/student/ssr": {
            "post": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Start SSR",
                "parameters": [
                    {
                        "description": "ssr info",
                        "name": "ApplyBid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSSR"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StViewRelation"
                        }
                    }
                }
            }
        },
        "/api/student/work": {
            "get": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "GetUserInfo student's works",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "student_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StWorks"
                        }
                    }
                }
            }
        },
        "/api/student/work/supervisor": {
            "get": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "GetUserInfo supervisors of the work",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Work ID",
                        "name": "work_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.WorkSvPlenty"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ApplyBid": {
            "type": "object",
            "properties": {
                "studentID": {
                    "type": "integer"
                },
                "supervisorID": {
                    "type": "integer"
                },
                "workID": {
                    "type": "integer"
                }
            }
        },
        "dto.ApplyBidResp": {
            "type": "object",
            "properties": {
                "bidID": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateSSR": {
            "type": "object",
            "properties": {
                "bidID": {
                    "type": "integer"
                },
                "studentID": {
                    "type": "integer"
                }
            }
        },
        "dto.FeedbackAddResp": {
            "type": "object",
            "properties": {
                "feedback_id": {
                    "type": "integer"
                }
            }
        },
        "dto.FeedbackPlenty": {
            "type": "object",
            "properties": {
                "feedbacks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FeedbackResp"
                    }
                }
            }
        },
        "dto.FeedbackReq": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "studentID": {
                    "type": "integer"
                },
                "supervisorID": {
                    "type": "integer"
                },
                "workID": {
                    "type": "integer"
                }
            }
        },
        "dto.FeedbackResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "studentID": {
                    "type": "integer"
                },
                "student_full_name": {
                    "type": "string"
                },
                "supervisorID": {
                    "type": "integer"
                },
                "workID": {
                    "type": "integer"
                },
                "work_kind": {
                    "type": "string"
                },
                "work_subject": {
                    "type": "string"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "dto.StBid": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "supervisor": {
                    "$ref": "#/definitions/dto.SvProfile"
                },
                "work": {
                    "$ref": "#/definitions/dto.Work"
                }
            }
        },
        "dto.StBids": {
            "type": "object",
            "properties": {
                "bids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StBid"
                    }
                }
            }
        },
        "dto.StProfile": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "studentCard": {
                    "type": "string"
                },
                "studentID": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "dto.StViewRelation": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "supervisor": {
                    "$ref": "#/definitions/dto.SvProfile"
                },
                "work": {
                    "$ref": "#/definitions/dto.Work"
                }
            }
        },
        "dto.StWork": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_started": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
        "dto.StWorks": {
            "type": "object",
            "properties": {
                "studentID": {
                    "type": "integer"
                },
                "works": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StWork"
                    }
                }
            }
        },
        "dto.SubjectResp": {
            "type": "object",
            "properties": {
                "department": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subjectID": {
                    "type": "integer"
                }
            }
        },
...
        "dto.SvProfile": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "avatarUrl": {
                    "type": "string"
                },
                "birthdate": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "supervisorID": {
                    "type": "integer"
                }
            }
        },
...
        "dto.Work": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "semester": {
                    "type": "integer"
                },
                "subject": {
                    "$ref": "#/definitions/dto.SubjectResp"
                }
            }
        },
        "dto.WorkSv": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "avatarUrl": {
                    "type": "string"
                },
                "birthdate": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "full": {
                    "type": "boolean"
                },
                "head": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "supervisorID": {
                    "type": "integer"
                }
            }
        },
        "dto.WorkSvPlenty": {
            "type": "object",
            "properties": {
                "supervisors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.WorkSv"
                    }
                },
                "workID": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Auth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}